ui/input: Use bool for qemu_add_mouse_event_handler()
Change kbd_mouse_{is,has}_absolute() alongside.

vmmouse uses uint8_t in its VMStateDescription, and Xen uses int in xend
communication, so they cannot easily be changed.

+++ hw/input/vmmouse.c
@@ -114,7 +114,7 @@ static void vmmouse_remove_handler(VMMouseState *s)
     }
 }
 
-static void vmmouse_update_handler(VMMouseState *s, int absolute)
+static void vmmouse_update_handler(VMMouseState *s, uint8_t absolute)
 {
     if (s->status != 0) {
         return;
